+++ sdk/ai/ai-agents/samples/v1-beta/javascript/agentsBingGroundingWithStreaming.js
@@ -72,7 +72,6 @@ async function main() {
           });
         }
         break;
-
       case RunStreamEvent.ThreadRunCompleted:
         console.log("Thread Run Completed");
         break;
@@ -89,19 +88,14 @@ async function main() {
   await client.deleteAgent(agent.id);
   console.log(`Deleted agent, agent ID: ${agent.id}`);
 
-  // Fetch and log all messages
-  console.log(`Messages:`);
   // Convert the PagedAsyncIterableIterator to an array of messages
-  const messagesArray = [];
   for await (const m of client.messages.list(thread.id)) {
-    messagesArray.push(m);
-  }
-
-  if (messagesArray.length > 0 && messagesArray[0].content.length > 0) {
-    const agentMessage = messagesArray[0].content[0];
-    if (isOutputOfType(agentMessage, "text")) {
-      const textContent = agentMessage;
-      console.log(`Text Message Content - ${textContent.text.value}`);
+    if (m.content.length > 0) {
+      const agentMessage = m.content[0];
+      if (isOutputOfType(agentMessage, "text")) {
+        console.log(`Text Message Content - ${agentMessage.text.value}`);
+      }
+      break; // Only log the first message content
     }
   }
 }