+++ sdk/ai/ai-agents/samples/v1-beta/javascript/agentsSharepoint.js
@@ -56,7 +56,7 @@ async function main() {
       intervalInMs: 2000,
     },
     onResponse: (response) => {
-      console.log(`Received response with status: ${response.status}`);
+      console.log(`Received response with status: ${response.parsedBody.status}`);
     },
   });
   console.log(`Run finished with status: ${run.status}`);
@@ -67,13 +67,11 @@ async function main() {
 
   // Fetch and log all messages
   const messagesIterator = await client.messages.list(thread.id);
-  console.log(`Messages:`);
   // Get the first message
   for await (const m of messagesIterator) {
     const agentMessage = m.content[0];
     if (isOutputOfType(agentMessage, "text")) {
-      const textContent = agentMessage;
-      console.log(`Text Message Content - ${textContent.text.value}`);
+      console.log(`Text Message Content - ${agentMessage.text.value}`);
     }
     break; // Only process the first message
   }