+++ sdk/ai/ai-agents/samples/v1-beta/javascript/batchVectorStoreWithFilesAndPolling.js
@@ -47,7 +47,7 @@ async function main() {
     fileIds: [file1.id, file2.id],
     // (Optional) Define an onResponse callback to monitor the progress of polling
     onResponse: (response) => {
-      console.log(`Received response with status: ${response.status}`);
+      console.log(`Received response with status: ${response.parsedBody.status}`);
     },
   });
   console.log(
@@ -61,7 +61,7 @@ async function main() {
   const vectorStoreFileBatchPoller = client.vectorStoreFileBatches.createAndPoll(vectorStore.id, {
     fileIds: [file1.id, file2.id],
     onResponse: (response) => {
-      console.log(`Received response with status: ${response.status}`);
+      console.log(`Received response with status: ${response.parsedBody.status}`);
     },
   });
   const vectorStoreFileBatch2 = await vectorStoreFileBatchPoller.pollUntilDone({