+++ sdk/avs/arm-avs/src/api/iscsiPaths/index.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export { $delete, createOrUpdate, get, listByPrivateCloud } from "./operations.js";
+export {
+  IscsiPathsDeleteOptionalParams,
+  IscsiPathsCreateOrUpdateOptionalParams,
+  IscsiPathsGetOptionalParams,
+  IscsiPathsListByPrivateCloudOptionalParams,
+} from "./options.js";