+++ sdk/healthinsights/health-insights-radiologyinsights-rest/CHANGELOG.md
@@ -1,7 +1,6 @@
 # Release History
 
-## 2.0.0 (Unreleased)
-
+## 2.0.0 (2025-06-06)
 Extending scope with inferences for scoring and assessment, quality measure and clinical guidance
 
 ### Features Added
@@ -40,7 +39,6 @@ Extending scope with inferences for scoring and assessment, quality measure and 
 ## 1.0.0 (2024-08-15)
 
  - GA release
- 
  
  ### Features Added
 - Added sample code for all Inference types